+++ apps/user/controller/Index.php
@@ -1,4 +1,13 @@
 <?php
+// 用户首页
+// +----------------------------------------------------------------------
+// | Copyright (c) 2016-2017 http://www.eacoo123.com, All rights reserved.         
+// +----------------------------------------------------------------------
+// | [EacooPHP] 并不是自由软件,可免费使用,未经许可不能去掉EacooPHP相关版权。
+// | 禁止在EacooPHP整体或任何部分基础上发展任何派生、修改或第三方版本用于重新分发
+// +----------------------------------------------------------------------
+// | Author:  心云间、凝听 <981248356@qq.com>
+// +----------------------------------------------------------------------
 namespace app\user\controller;
 use app\home\controller\Home;
 
@@ -17,9 +26,10 @@ class Index extends Home{
      * */
     public function index(){
 
-        $map['status'] = ['egt', '0']; // 禁用和正常状态
+        $map['status'] = 1; // 禁用和正常状态
         list($user_list) = $this->userModel->getListByPage($map,'reg_time desc','*',20);
         $this->assign('user_list',$user_list);
+        $this->pageConfig('会员列表','users');
         return $this->fetch();
 
     }
@@ -29,63 +39,19 @@ class Index extends Home{
      *  By: yyyvy  <QQ:76836785>
      *  Time: 2017-12-28 06:55:11
      * */
-    public function home($uid){
-        $info = userModel::info($uid);
-        $this->assign('info',$info);
-        return $this->fetch();
-    }
-
-    /*
-     *  Description: 会员登录
-     *  By: yyyvy  <QQ:76836785>
-     *  Time: 2017-12-28 10:19:55
-     * */
-    public function login(){
-        if (IS_POST) {
-            $data = input('post.');
-            $result = $this->validate($data, [
-              ['username', 'require|min:1', '登录名不能为空|登录名格式不正确'],
-              ['password', 'require|length:6,32', '请填写密码|密码格式不正确']
-            ]);
-            if (true !== $result) {
-                // 验证失败 输出错误信息
-
-                $this->error($result);
-
-                exit;
+    public function home($uid = 0){
+        try {
+            if ($uid>0) {
+                $info = userModel::info($uid);
+                $this->assign('info',$info);
+                return $this->fetch();
             }
-            if(isset($data['rememberme'])){
-                $rememberme = $data['rememberme']==1 ? true : false;
-            }else{
-                $rememberme = false;
-            }
-
-            $result = UserModel::login($data['username'],$data['password'], $rememberme);
-            //print_r($result);die;
-
-            if ($result['code']==1) {
-
-                $uid = !empty($result['data']['uid']) ? $result['data']['uid']:0;
-                $this->success('登录成功！',url('/'));
-            } elseif ($result['code']==0) {
-                $this->error($result['msg']);
-            } else {
-                $this->logout();
-            }
-        }else{
-            return $this->fetch();
+            throw new \Exception("用户ID不正确", 0);
+            
+        } catch (\Exception $e) {
+            $this->error($e->getMessage());
         }
-    }
-
-    /*
-     *  Description: 退出登录
-     *  By: yyyvy  <QQ:76836785>
-     *  Time: 2017-12-28 10:19:55
-     * */
-    public function logout(){
-        session(null);
-        cookie(null,config('cookie.prefix'));
-        $this->redirect('/');
+        
     }
 
     /*
