+++ public/themes/default/pc/public/layout.html
@@ -35,6 +35,7 @@
 <script src="https://oss.maxcdn.com/html5shiv/3.7.3/html5shiv.min.js"></script>
 <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
 <![endif]-->
+{:hook('PageHeader')}
 </head>
 <body class="hold-transition skin-blue layout-top-nav">
  <div class="wrapper">
@@ -172,7 +173,7 @@
             </li>
 
             {empty name="current_user"}
-              <li><a href="{:url('user/index/login')}">登录</a></li>
+              <li><a href="{:url('user/login/login')}">登录</a></li>
             {else/}
               <li class="dropdown user user-menu">
               <!-- Menu Toggle Button -->
@@ -212,7 +213,7 @@
                     <a href="{:url('user/index/personal')}" class="btn btn-default btn-flat">个人资料</a>
                   </div>
                   <div class="pull-right">
-                    <a href="{:url('user/index/logout')}" class="btn btn-default btn-flat">退出</a>
+                    <a href="{:url('user/login/logout')}" class="btn btn-default btn-flat">退出</a>
                   </div>
                 </li>
               </ul>
@@ -245,7 +246,7 @@
 <script type="text/javascript" src="__STATIC__/libs/fastclick/fastclick.min.js"></script>
 <!-- AdminLTE App -->
 <script type="text/javascript" src="__ADMIN_JS__/app.min.js" ></script>
-
+  {:hook('PageFooter')}
 {:config('web_site_statistics')}
 {block name="script"}{/block}
