+++ public/themes/default/pc/user/index/home.html
@@ -21,6 +21,7 @@
                         </div>
                         <div class="p-username">
                             <a href=""><strong>{$info.nickname}</strong></a>
+                            {:hook('attention',['type'=>'attadd','uid'=>$info['uid']])}
                         </div>
                         <div class="p-exchange">
                             <span class="pipe"></span>
@@ -33,6 +34,7 @@
                             <i class="fa fa-leaf" style="margin-right: 5px;"></i>
                             最后登录：{$info.last_login_time,'ymd'|friendly_date}
                         </div>
+
                         <div class="alert alert-warning" style="width: 50%;margin: 20px auto;">签名：{$info.description}</div>
                     </div>
                     <div class="p-navs">
