+++ public/themes/default/pc/user/index/index.html
@@ -7,6 +7,11 @@
 
 <div class="content-wrapper">
     <div class="container">
+        <section class="content-header">
+        <ol class="breadcrumb">
+          {$page_config.breadcrumbs|default=''}
+        </ol>
+      </section>
         <!-- Main content -->
         <section class="content">
             <div class="row">
