--- public/themes/default/pc/user/index/login.html
@@ -1,70 +1,0 @@
-{extend name="$_theme_public_layout"/}
-{block name="style"}
-
-{/block}
-
-{block name="main"}
-
-<div class="content-wrapper">
-    <div class="container">
-
-        <!-- Main content -->
-        <section class="content">
-            <div class="row">
-                <div class="login">
-                    <form action="{:url('index/login')}" method="post" class="container offset1 loginform">
-                        <div id="owl-login">
-                            <div class="hand"></div>
-                            <div class="hand hand-r"></div>
-                            <div class="arms">
-                                <div class="arm"></div>
-                                <div class="arm arm-r"></div>
-                            </div>
-                        </div>
-                        <div class="pad">
-                            <div class="control-group">
-                                <div class="controls">
-                                    <label for="username" class="control-label fa fa-user"></label>
-                                    <input id="username" type="username" name="username" placeholder="账号/邮箱" tabindex="1" autofocus="autofocus" class="form-control input-medium">
-                                </div>
-                            </div>
-                            <div class="control-group">
-                                <div class="controls">
-                                    <label for="password" class="control-label fa fa-asterisk"></label>
-                                    <input id="password" type="password" name="password" placeholder="密码" tabindex="2" class="form-control input-medium">
-                                </div>
-                            </div>
-                            <div class="control-group">
-                                <div class="controls">
-                                    <input type="checkbox" id="rememberme" name="rememberme" value="0">
-                                    <span></span> 7天内自动登录?
-                                </div>
-                            </div>
-                        </div>
-                        <div class="form-actions">
-                            <a href="#" tabindex="5" class="btn pull-left btn-link text-muted">忘记密码?</a>
-                            <a href="#" tabindex="6" class="btn btn-link text-muted">立即注册</a>
-                            <button type="submit" tabindex="4" class="btn btn-primary">登录</button>
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </section>
-    </div>
-</div>
-<!-- /.content-wrapper -->
-
-{/block}
-
-{block name="script"}
-<script type="application/javascript">
-  $(function() {
-    $('.login #password').focus(function() {
-      $('#owl-login').addClass('password');
-    }).blur(function() {
-      $('#owl-login').removeClass('password');
-    });
-  });
-</script>
-
-{/block}
